Improved codegen scripts to work under gopath

+++ scripts/verify-codegen.sh
@@ -18,30 +18,40 @@ set -o errexit
 set -o nounset
 set -o pipefail
 
-SCRIPT_ROOT=$(dirname "${BASH_SOURCE[0]}")/..
-DIFFROOT="${GOPATH}/src/github.com/skupperproject/skupper/pkg/generated"
-TMP_DIFFROOT="${SCRIPT_ROOT}/_tmp/pkg"
+SCRIPT_ROOT=$(cd `dirname "${BASH_SOURCE[0]}"`/.. && pwd)
+DIFFROOT_PATHS=("/pkg/apis" "/pkg/generated")
 _tmp="${SCRIPT_ROOT}/_tmp"
 
 cleanup() {
-  rm -rf "${_tmp}"
+  # sanity check
+  if [[ -f ${SCRIPT_ROOT}/go.mod ]] && grep -q '^module github.com/skupperproject/skupper$' ${SCRIPT_ROOT}/go.mod; then
+    rm -rf "${_tmp}"
+  fi
 }
 trap "cleanup" EXIT SIGINT
 
 cleanup
 
-mkdir -p "${TMP_DIFFROOT}"
-cp -a ${DIFFROOT} "${TMP_DIFFROOT}"
+# adding current files
+mkdir -p "${_tmp}/pkg"
+for diffroot_path in ${DIFFROOT_PATHS[@]}; do
+    cp -r ${SCRIPT_ROOT}${diffroot_path} ${_tmp}/pkg/
+done
 
-DO_NOT_UPDATE=true "${SCRIPT_ROOT}/scripts/update-codegen.sh"
-echo "diffing ${DIFFROOT} against freshly generated codegen"
-ret=0
-diff -Naupr "${DIFFROOT}" "${TMP_DIFFROOT}/generated" || ret=$?
-cp -a "${TMP_DIFFROOT}"/generated/* "${DIFFROOT}/"
-if [[ $ret -eq 0 ]]
-then
-  echo "${DIFFROOT} up to date."
-else
-  echo "${DIFFROOT} is out of date. Please run scripts/update-codegen.sh"
-  exit 1
+DO_NOT_UPDATE=true "${SCRIPT_ROOT}/scripts/update-codegen.sh" "${_tmp}"
+OUTDATED_PATHS=()
+echo "comparing against freshly generated codegen"
+for diffroot_path in ${DIFFROOT_PATHS[@]}; do
+  ret=0
+  diff -Naupr "${SCRIPT_ROOT}${diffroot_path}" "${_tmp}${diffroot_path}" || ret=$?
+  if [[ $ret -ne 0 ]]
+  then
+    OUTDATED_PATHS+=(${diffroot_path})
+  fi
+done
+
+if [[ ${#OUTDATED_PATHS[@]} -gt 0 ]]; then
+    echo "${OUTDATED_PATHS[@]} out of date. Please run scripts/update-codegen.sh"
+    exit 1
 fi
+echo "${DIFFROOT_PATHS[@]} up to date."
